+++ src/analysis.cpp
@@ -13,14 +13,16 @@ std::string Analysis::asJson() {
   std::stringstream ss;
   ss << "{" << std::endl;
 
-  ss << JSTR(0,"slippi_version",   slippi_version)  << ",\n";
-  ss << JSTR(0,"parser_version",   parser_version)  << ",\n";
-  ss << JSTR(0,"analyzer_version", analyzer_version)<< ",\n";
-  ss << JSTR(0,"game_time",        game_time)       << ",\n";
-  ss << JUIN(0,"stage_id",         stage_id)        << ",\n";
-  ss << JSTR(0,"stage_name",       stage_name)      << ",\n";
-  ss << JUIN(0,"game_length",      game_length)     << ",\n";
-  ss << JUIN(0,"winner_port",      winner_port)     << ",\n";
+  ss << JSTR(0,"original_file",    escape_json(original_file))  << ",\n";
+  ss << JSTR(0,"slippi_version",   slippi_version)              << ",\n";
+  ss << JSTR(0,"parser_version",   parser_version)              << ",\n";
+  ss << JSTR(0,"analyzer_version", analyzer_version)            << ",\n";
+  ss << JSTR(0,"game_time",        game_time)                   << ",\n";
+  ss << JUIN(0,"stage_id",         stage_id)                    << ",\n";
+  ss << JSTR(0,"stage_name",       stage_name)                  << ",\n";
+  ss << JUIN(0,"game_length",      game_length)                 << ",\n";
+  ss << JUIN(0,"winner_port",      winner_port)                 << ",\n";
+  ss << JUIN(0,"start_minutes",    timer)                       << ",\n";
 
   ss << "\"players\" : [\n";
   for(unsigned p = 0; p < 2; ++p) {
@@ -96,6 +98,24 @@ std::string Analysis::asJson() {
     ss << JFLT(1,"actions_per_min",        ap[p].apm)                       << ",\n";
     ss << JUIN(1,"state_changes",          ap[p].state_changes)             << ",\n";
     ss << JFLT(1,"states_per_min",         ap[p].aspm)                      << ",\n";
+    ss << JUIN(1,"shieldstun_times",       ap[p].shieldstun_times)          << ",\n";
+    ss << JUIN(1,"shieldstun_act_frames",  ap[p].shieldstun_act_frames)     << ",\n";
+    ss << JUIN(1,"hitstun_times",          ap[p].hitstun_times)             << ",\n";
+    ss << JUIN(1,"hitstun_act_frames",     ap[p].hitstun_act_frames)        << ",\n";
+    ss << JUIN(1,"wait_times",             ap[p].wait_times)                << ",\n";
+    ss << JUIN(1,"wait_act_frames",        ap[p].wait_act_frames)           << ",\n";
+    ss << JUIN(1,"used_norm_moves",        ap[p].used_norm_moves)           << ",\n";
+    ss << JUIN(1,"used_spec_moves",        ap[p].used_spec_moves)           << ",\n";
+    ss << JUIN(1,"used_misc_moves",        ap[p].used_misc_moves)           << ",\n";
+    ss << JUIN(1,"used_grabs",             ap[p].used_grabs)                << ",\n";
+    ss << JUIN(1,"used_pummels",           ap[p].used_pummels)              << ",\n";
+    ss << JUIN(1,"used_throws",            ap[p].used_throws)               << ",\n";
+    ss << JUIN(1,"total_moves_used",       ap[p].total_moves_used)          << ",\n";
+    ss << JUIN(1,"total_moves_landed",     ap[p].total_moves_landed)        << ",\n";
+    ss << JFLT(1,"move_accuracy",          ap[p].move_accuracy)             << ",\n";
+    ss << JFLT(1,"actionability",          ap[p].actionability)             << ",\n";
+    ss << JFLT(1,"neutral_wins_per_min",   ap[p].neutral_wins_per_min)      << ",\n";
+    ss << JFLT(1,"mean_death_percent",     ap[p].mean_death_percent)        << ",\n";
 
     ss << SPACE[ILEV] << "\"interaction_frames\" : {\n";
     for(unsigned d = Dynamic::__LAST-1; d > 0; --d) {
@@ -111,7 +131,7 @@ std::string Analysis::asJson() {
 
     ss << SPACE[ILEV] << "\"moves_landed\" : {\n";
     unsigned _total_moves = 0;
-    for(unsigned d = 0; d < Move::__LAST; ++d) {
+    for(unsigned d = 0; d < Move::BUBBLE; ++d) {
       if ((ap[p].move_counts[d]) > 0) {
         ss << JUIN(2,Move::name[d], ap[p].move_counts[d]) << ",\n";
         _total_moves += ap[p].move_counts[d];
@@ -121,13 +141,6 @@ std::string Analysis::asJson() {
     ss << SPACE[ILEV] << "},\n";
 
     ss << SPACE[ILEV] << "\"attacks\" : [\n";
-    // move_id
-    // anim_frame
-    // punish_id
-    // frame
-    // damage
-    // opening
-    // kill_dir
     for(unsigned i = 0; ap[p].attacks[i].frame > 0; ++i) {
       ss << SPACE[2*ILEV] << "{" << std::endl;
       ss << JUIN(2,"move_id",         ap[p].attacks[i].move_id)                        << ",\n";